### 2025-09-22 #9054 @danielskatz
👋 @cucuwritescode - can you explain what part of the code you have submitted in this repository/branch is to be reviewed by JOSS? Is there a PR, branch, diff, etc that can be isolated for this?
