### 2025-09-22 #9054 @cucuwritescode
Hi @danielskatz, thanks for picking this up!

To clarify, the JOSS submission covers the `faust2clap` tool only. In this repository/branch (`faust2clap-joss`), the relevant code is contained in the following:

- `tools/faust2clap/faust2clap.py` — static build script
- `architecture/clap/*` — CLAP architecture files, including `clap-arch.cpp`, `CLAPMapUI`, and dynamic hot-reload support
- Supporting files for the dynamic hot-reload plugin: `faust-hot-reload.py`, `Makefile.simple`

---

Everything else in this fork comes from the upstream Faust repository. For the purpose of JOSS review, only the above files are to be considered.
