### 2025-09-22 #9054 @danielskatz
👋 @cucuwritescode - where are the latter two files?

And it looks like all of this was committed in the last month, correct? If so, can you say anything about the amount of work you did, as this seems like it would be less scholarly effort than JOSS would normally accept for review.
