### 2025-10-07 #9054 @danielskatz
👋 @cucuwritescode - your submission has passed our scope review.
