### 2025-11-18 #9054 @JorenSix
@diehlpk I am a bit hesitant to review this work since I am not too familiar with the Faust ecosystem but I am happy to help and review.
