### 2025-11-24 #9508 @editorialbot
[REVIEW]: faust2clap: Generating CLAP Plug-ins from Faust DSP Code
**Submitting author:** @cucuwritescode (<a href="http://orcid.org/0009-0008-6654-0401">Facundo Franchino</a>)
**Repository:** https://github.com/cucuwritescode/faust2clap
**Branch with paper.md** (empty if default branch): faust2clap-joss
**Version:** v1.0.0
