### 2026-02-02 #9508 @cucuwritescode
Hello! I’m sorry for the delay, I was in the midst of my last round of undergrad finals plus work.
Commited yesterday with the suggested changes. I modified the paper and repo based on the issues and general feedback on this thread. I’m afraid I can’t say academics have used faust2clap with concrete evidence, yet.
Though it’s worth mentioning it’s part of the Faust Programming Language, thence it’s widely available to a few thousands of people.
On a sidenote, I am planning to use it in a (yet unpublished/ submitted) paper for DAFx ’26, and I've used it to automatically generate plugins and “live code” my reverb algorithms in this publicly available article I wrote: https://facundofranchino.com/#reader/reverberation/emulating-space
I’ve also exchanged emails with Julius Orion Smith and I’m aware he’s tried it, but I’m unsure whether it’ll be part of his research at some point.

Thank you, hope this clears up a few things.
