### 2026-02-02 #9508 @diehlpk
@cucuwritescode I have some fundamental questions. It seems you forked the original faust code and added some plugin.

Is that correct? If so how will you sync your fork with the faust if they release new versions?
