### 2026-02-02 #9508 @cucuwritescode
Hello, I forked faust's repo and added a compiler architecture for the CLAP plugin standard, called faust2clap.
For new versions of faust2clap I can simply open a new PR, as I did with the versions that got merged to the language post-Google Summer of Code. By doing this, one of the official maintainers (mainly my supervisor for this project Stéphane Letz), will run the new faust2clap version through automated tests/manual checks within the larger codebase of the language to make sure everything's running smoothly.
This is the most practical approach and "best practice", given Faust's large codebase and limited number of maintainers.
Hope that answers your question.
Best,
Facundo
