### 2026-02-03 #9508 @diehlpk
@cucuwritescode Can you isolate the code from your plugin? So the reviewers would only review your code and not faust?

Should your plugin be added to [this list](https://faustdoc.grame.fr/manual/tools/#faust2-tools)?
