### 2026-02-03 #9508 @diehlpk
@cucuwritescode Do you have tests for your plugin?

Do you have documentation for your plugin?
