### 2026-02-11 #9508 @cucuwritescode
> [@cucuwritescode](https://github.com/cucuwritescode) Can you isolate the code from your plugin? So the reviewers would only review your code and not faust?
>
> Should your plugin be added to [this list](https://faustdoc.grame.fr/manual/tools/#faust2-tools)?

### Code location
  My contributions are isolated to:
  - `tools/faust2clap/` - the build tool/CLI etc
  - `architecture/clap/` - the CLAP architecture files
Yes, it should. But that hasn’t been updated in a while by the main maintainers
