### 2026-02-11 #9508 @cucuwritescode
> [@cucuwritescode](https://github.com/cucuwritescode) Do you have tests for your plugin?
>
> Do you have documentation for your plugin?

### Tests
The tool includes manual testing documentation (see README "Testing" section). Automated CI tests are on the roadmap but not yet implemented, the tool is tested manually against various Faust DSP examples. There was testing done by the main maintainer of the language’s codebase prior to merging the contributions

### Documentation
 Documentation is in:
  - `tools/faust2clap/README.md`
  - `README.md` (root)
