### 2026-02-11 #9508 @cucuwritescode
> > [@diehlpk](https://github.com/diehlpk) Next to the question about which branch I am also unclear where the contributions of the author start and stop. Since the author developed a part of the faust ecosystem it is difficult to evaluate. For example: does the software have automated tests: the main faust package has a number of tests, the faust2clap tool does not (or I am looking in the wrong place).
>
> [@JorenSix](https://github.com/JorenSix) see my above questions.

### Branch
  The code for review is on the `faust2clap-joss` branch.
