### 2026-02-11 #9508 @cucuwritescode
@diehlpk A bit more but roughly speaking yes
