### 2026-02-11 #9508 @diehlpk
@JorenSix, @yiitozer

Please only review the files in these folders for this review

tools/faust2clap/ - the build tool/CLI etc
architecture/clap/ - the CLAP architecture files
