### 2026-02-11 #9508 @diehlpk
@cucuwritescode

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

I could not see any documentation in your code.
