### 2026-02-11 #9508 @cucuwritescode
The code contains inline documentation throughout:
  - Python functions have docstrings explaining their purpose
  - C++ files have inline comments explaining the CLAP integration logic

  Usage documentation is in the READMEs. Since faust2clap is a CLI tool (not a
  library), there's no public API to document in the traditional sense, users
  interact via command-line flags which are documented in `--help` and the
  README.

  Happy to add more inline comments if specific areas are unclear.
